+++ CHANGES.rst
@@ -36,6 +36,12 @@ heasarc
 - Add alternative instance of HEASARC Server, maintained by
   INTEGRAL Science Data Center. [#1988]
 
+irsa
+^^^^
+
+- Adding ``cache`` kwarg to the class methods to be able to control the use
+  of local cache. [#2092]
+
 nasa_exoplanet_archive
 ^^^^^^^^^^^^^^^^^^^^^^